requirements.txt
aiohttp==3.8.3
arrow==0.15.5
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bleach==3.3.0
certifi==2020.4.5.1
cffi==1.14.5
chardet==3.0.4
click==7.1.1
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
geojson==2.5.0
idna==2.9
ipykernel==5.2.1
ipython==7.16.3
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
MouseInfo==0.1.3
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.4.10
packaging==20.9
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
ptyprocess==0.6.0
PyAutoGUI==0.9.50
pycparser==2.20
PyGetWindow==0.0.8
Pygments==2.7.4
pylint==2.4.4
PyMsgBox==1.0.8
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
pyrsistent==0.16.0
PyScreeze==0.1.26
python-dateutil==2.8.1
python-dotenv==0.13.0
PyTweening==1.0.3
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
requests==2.25.1
six==1.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.1
traitlets==4.3.3
urllib3==1.26.5
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.11.2
zipp==3.1.0
redis~=4.3.4
Deprecated~=1.2.13
Send2Trash~=1.8.0
netifaces~=0.11.0